[refactor] Engine 리팩토링
